+++ src/js/leaflet.storage.js
@@ -877,6 +877,7 @@ L.Storage.Map.include({
             'captionBar',
             'slideshow',
             'sortKey',
+            'filterKey',
             'showLabel',
             'shortCredit',
             'longCredit'
@@ -1024,7 +1025,8 @@ L.Storage.Map.include({
             'options.popupContentTemplate',
             'options.zoomTo',
             'options.showLabel',
-            ['options.sortKey', {handler: 'Input', helpText: L._('Property to use for sorting features'), placeholder: L._('Default: name')}]
+            ['options.sortKey', {handler: 'Input', helpText: L._('Property to use for sorting features'), placeholder: L._('Default: name')}],
+            ['options.filterKey', {handler: 'Input', helpText: L._('Comma separated list of properties to use when filtering features'), placeholder: L._('Default: name')}]
         ];
 
         builder = new L.S.FormBuilder(this, optionsFields, {